[PB-2408] refactor create team
